NMS-12863: Separated Events unit tests from XML tests
